javascript evaluation ios/android (#1021)
Create evaluateJavascript method in the WebView plugin. 

As a result of updating the enum name from JavaScriptMode to JavascriptMode, which is a breaking change, the version number has been updated to 0.2.0.
Android and IOS return different types for javascript evaluation result, we unified them to make the result type to be string. 
In some ObjectiveC code, change == [NSNull null] to isKindOfClass:[NSNull class] to avoid type warning.