[IAP] Generate boilerplate serializers (#1090)
Brings in `json_serializable` as a build step to generate the Dart
platform channel parsing boilerplate. This dependency is still very
experimental and may be abandoned if the extra build step ends up being
a maintenance burden greater than maintaining the boilerplate. At this
point I expect it to be worth the trouble given the sheer amount of data
classes and deserialization that's going to happen in this plugin.

Also converts `SkuType` and `BillingResponse` to enums.