[In_app_purchase]fix requesthandler crash (#1199)
The FIAPRequestHandler is destroyed when exiting the function. Since FIAPRequestHandler is the delegate, when SKProductRequest tries to execute the delegate method, it will not be able to find the delegate. Hence the crash.

Adding a set to hold strong reference of the FIAPRequestHandler until it is complete fix the crash.