[In_App_Purchase] Improve testability (#2016)
This PR creates a factory for BillingClient, with that was possible to remove a secondary constructor that was used just for test.

With the introduction of the factory there was the need to update some tests as it was testing states that would never be reached in the normal plugin flow (due the usage of the constructor for testing). Basically, the test cases depended on the state of BillingClient that would be null if no previous setup was executed.