[url_launcher] Document why canLaunch can return false on Android API 30 (#3040)
This PR adds extra documentation to the canLaunch method, explaining that (from Android API 30) false is returned when the package visibility is not managed properly.

Starting from Android API 30, apps should explicitly list which other apps it wants to query or interact with, by specifying them in the AndroidManifest.xml file. More information can be found in the Manage package visibility article in the Android documentation.